refactor: cleaned up the code and added the profile section to a helper component
